compress js/css files with yuicompressor one-by-one on failure
Yuicompressor may fail on some unsupported language construct, i.e.:

var foo = {}; foo.in = 'bar';

will throw an error ("in" is also a js keyword)- but maybe it is valid to use it that way.
If the compression fails on a single file, we run yuicompressor for each source-file separate, and decide (on error) to append the compressed or original file to the outfile.