add policy to DBFilterGroup
There is a „save current filters“ in SESSION on the ToDo list.

Although the current modification seems to allow it, it is not perfect, because it only saves filter values for a single Url - ideally it should grab the current action, so /foo/bar/4/list and /foo/bar/6/list could share saved filter values.

But how to achieve it? Feel free to comment or improve.